V-02 Decomposition of the navigation module
